refactor: dataCount改为Long型

+++ src/main/java/tech/zhiwei/frostmetal/modules/mydata/manage/entity/Data.java
@@ -37,6 +37,6 @@ public class Data extends TenantEntity {
     /**
      * 业务数据
      */
-    private Integer dataCount;
+    private Long dataCount;
 
 }
